create separate singleton classes for comlink workers
