+++ src/utils/comlink/ComlinkConvertWorker.ts
@@ -1,0 +1,30 @@
+import { Remote } from 'comlink';
+import { runningInBrowser } from 'utils/common';
+import { DedicatedConvertWorker } from 'worker/convert.worker';
+import { ComlinkWorker } from './comlinkWorker';
+
+class ComlinkConvertWorker {
+    private comlinkWorkerInstance: Remote<DedicatedConvertWorker>;
+
+    async getInstance() {
+        if (!this.comlinkWorkerInstance) {
+            const comlinkWorker = getDedicatedConvertWorker();
+            this.comlinkWorkerInstance = await new comlinkWorker.remote();
+        }
+        return this.comlinkWorkerInstance;
+    }
+}
+
+export const getDedicatedConvertWorker = () => {
+    if (runningInBrowser()) {
+        const cryptoComlinkWorker = new ComlinkWorker<
+            typeof DedicatedConvertWorker
+        >(
+            'ente-convert-worker',
+            new Worker(new URL('worker/convert.worker.ts', import.meta.url))
+        );
+        return cryptoComlinkWorker;
+    }
+};
+
+export default new ComlinkConvertWorker();
